refactor: ticket 클래스 파일 model로 이동
