--- src/js/lotto/model.js
@@ -1,24 +1,0 @@
-class LottoModel {
-  #lottos
-
-  constructor() {
-    this.#lottos = []
-  }
-
-  get lottos() {
-    return this.#lottos
-  }
-
-  get amount() {
-    return this.#lottos.length
-  }
-
-  addLotto(newLotto) {
-    this.#lottos.push(newLotto)
-  }
-
-  init() {
-    this.#lottos = []
-  }
-}
-export default LottoModel
